Change fibonacci data types to long

+++ src/tests/benchmarks/Fibonacci.cc
@@ -9,7 +9,7 @@
 #include <stdio.h>
 #include <omp.h>
 
-int * memoizer;
+long * memoizer;
 long comp_fib_numbers(int n) {
   if ( n == 0 || n == 1 ) return n;
 
@@ -24,7 +24,7 @@ long comp_fib_numbers(int n) {
   }
 
   #pragma omp taskwait
-  int fn = memoizer[n - 1] + memoizer[n - 2];
+  long fn = memoizer[n - 1] + memoizer[n - 2];
 
   return fn;
 }
@@ -38,7 +38,7 @@ int main(int argc, char **argv) {
   }
   n = atoi(argv[1]);
 
-  memoizer = new int[n+1]();
+  memoizer = new long[n+1]();
 //#pragma omp parallel num_threads(NUM_THREADS)
 #pragma omp parallel num_threads(2)
   {
